[3단계 - 테마 확장 기능] refactor: reservation 도메인 패키지 이동
