+++ src/main/java/nextstep/application/reservation/RoomEscapeService.java
@@ -1,14 +1,14 @@
-package nextstep.application;
+package nextstep.application.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import nextstep.application.dto.Reservation;
-import nextstep.application.dto.ReservationRes;
-import nextstep.domain.ReservationEntity;
-import nextstep.domain.repository.ReservationRepository;
+import nextstep.application.reservation.dto.Reservation;
+import nextstep.application.reservation.dto.ReservationRes;
+import nextstep.domain.reservation.ReservationEntity;
+import nextstep.domain.reservation.repository.ReservationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
