+++ src/main/java/nextstep/config/DatabaseDriverConfig.java
@@ -1,7 +1,7 @@
 package nextstep.config;
 
-import nextstep.domain.repository.JdbcReservationRepository;
-import nextstep.domain.repository.ReservationRepository;
+import nextstep.domain.reservation.repository.JdbcReservationRepository;
+import nextstep.domain.reservation.repository.ReservationRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.core.JdbcTemplate;
