+++ src/main/java/nextstep/domain/reservation/ReservationEntity.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
